--- a/tempodb/encoding/vparquet3/block_traceql.go
+++ b/tempodb/encoding/vparquet3/block_traceql.go
@@ -1568,11 +1568,9 @@ func (c *spanCollector) String() string {
 func (c *spanCollector) KeepGroup(res *parquetquery.IteratorResult) bool {
 	var sp *span
 	// look for existing span first. this occurs on the second pass
-	for _, e := range res.OtherEntries {
-		if e.Key == otherEntrySpanKey {
-			sp = e.Value.(*span)
-			break
-		}
+	val := res.OtherValueFromKey(otherEntrySpanKey)
+	if val != nil {
+		sp = val.(*span)
 	}
 
 	// if not found create a new one
@@ -1667,8 +1665,7 @@ func (c *spanCollector) KeepGroup(res *parquetquery.IteratorResult) bool {
 		}
 	}
 
-	res.Entries = res.Entries[:0]
-	res.OtherEntries = res.OtherEntries[:0]
+	res.Reset()
 	res.AppendOtherValue(otherEntrySpanKey, sp)
 
 	return true
@@ -1772,8 +1769,7 @@ func (c *batchCollector) KeepGroup(res *parquetquery.IteratorResult) bool {
 		return false
 	}
 
-	res.Entries = res.Entries[:0]
-	res.OtherEntries = res.OtherEntries[:0]
+	res.Reset()
 	res.AppendOtherValue(otherEntrySpansetKey, sp)
 
 	return true
@@ -1849,8 +1845,7 @@ func (c *traceCollector) KeepGroup(res *parquetquery.IteratorResult) bool {
 		}
 	}
 
-	res.Entries = res.Entries[:0]
-	res.OtherEntries = res.OtherEntries[:0]
+	res.Reset()
 	res.AppendOtherValue(otherEntrySpansetKey, finalSpanset)
 
 	return true
@@ -1892,9 +1887,8 @@ func (c *attributeCollector) KeepGroup(res *parquetquery.IteratorResult) bool {
 		}
 	}
 
-	res.Entries = res.Entries[:0]
-	res.OtherEntries = res.OtherEntries[:0]
-	res.AppendOtherValue(key, val)
+	res.Reset()
+	res.AppendOtherValue(key, val) // jpe added here
 
 	return true
 }