--- a/tempodb/encoding/vparquet3/block_traceql.go
+++ b/tempodb/encoding/vparquet3/block_traceql.go
@@ -1884,7 +1884,8 @@ func (c *attributeCollector) KeepGroup(res *parquetquery.IteratorResult) bool {
 		}
 	}
 
-	if val.Type == traceql.TypeNil {
+	// if this row didn't match up a key/value pair then just drop it
+	if val.Type == traceql.TypeNil || key == "" {
 		return false
 	}