@joe-elliott 2023-09-01
Interestingly there was a regression on:
```
BackendBlockTraceQL/mixedValNoMatch-8                  345ms ± 1%      450ms ± 1%   +30.26%  (p=0.008 n=5+5)
```

Dumping two cpu profiles:
Before:
![image](https://github.com/grafana/tempo/assets/2272392/3a92581f-b09f-4a20-98fe-b5324731b819)

After:
![image](https://github.com/grafana/tempo/assets/2272392/4f0a6f4c-1428-4dc4-bbed-984a1e0960e9)

Note the additional time in `seekAll`. It makes sense that we are calling seekAll more since we are now skipping rows at the attribute level, but surprising (on the surface) that this appears to be less efficient then just iterating through all the values.